Move TransactionDisconnecting tests into the Connection namespace
