+++ tests/Database/Functional/Driver/MySQL/Connection/TransactionDisconnectingTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Database\Tests\Functional\Driver\MySQL\Query;
+namespace Cycle\Database\Tests\Functional\Driver\MySQL\Connection;
 
 // phpcs:ignore
-use Cycle\Database\Tests\Functional\Driver\Common\Query\TransactionDisconnectingTest as CommonClass;
+use Cycle\Database\Tests\Functional\Driver\Common\Connection\TransactionDisconnectingTest as CommonClass;
 
 /**
  * @group driver
