+++ tests/Database/Functional/Driver/SQLite/Connection/TransactionDisconnectingTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Database\Tests\Functional\Driver\SQLite\Query;
+namespace Cycle\Database\Tests\Functional\Driver\SQLite\Connection;
 
 // phpcs:ignore
-use Cycle\Database\Tests\Functional\Driver\Common\Query\TransactionDisconnectingTest as CommonClass;
+use Cycle\Database\Tests\Functional\Driver\Common\Connection\TransactionDisconnectingTest as CommonClass;
 use Cycle\Database\Tests\Stub\Driver\SQLiteWrapDriver;
 
 /**
